CARDS-2870: Reorganize styles into domain-specific modules Unused styles cleanup Undefined styles cleanup Fixed compilation error Reverted filter display name removal
