+++ modules/homepage/src/main/frontend/src/themePage/Sidebar/sidebarStyles.jsx
@@ -355,8 +355,6 @@ const sidebarStyles = theme => ({
     flexDirection: "column",
     justifyContent: "space-between"
   },
-  adminButton: {
-  },
   adminSidebar: {
     [theme.breakpoints.up("md")]: {
       width: "100%",
